little maj

+++ srcs/prs/ft_prschrs.c
@@ -6,7 +6,7 @@
 /*   By: glouyot <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/12/29 10:59:28 by glouyot           #+#    #+#             */
-/*   Updated: 2017/01/19 15:28:05 by glouyot          ###   ########.fr       */
+/*   Updated: 2017/01/19 15:41:56 by glouyot          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,7 +20,7 @@ t_byte		ft_prschrs(char **cursor, char const *chrs)
 
 	counts = 0;
 	curs = *cursor;
-	if (chrs && cursor && ft_updiv(ft_strlen(chrs), 8) <= sizeof(t_byte))
+	if (chrs && cursor && ft_updiv(ft_strlen(chrs), 8) <= (int)sizeof(t_byte))
 	{
 		while ((fonds = ft_strchr(chrs, *curs)))
 		{
